update to 1.20
