+++ build.gradle.kts
@@ -1,6 +1,9 @@
 plugins {
     id("paper-plugin")
-    //id("library-plugin")
+
+    //id("publish-task")
+
+    id("xyz.jpenilla.run-paper") version "2.1.0"
 }
 
 dependencies {
