+++ buildSrc/build.gradle.kts
@@ -5,7 +5,8 @@ plugins {
 repositories {
     gradlePluginPortal()
 
-    maven("https://repo.crazycrew.us/api/")
+    maven("https://repo.crazycrew.us/first-party/")
+    maven("https://repo.crazycrew.us/third-party/")
 }
 
 dependencies {
