+++ buildSrc/settings.gradle.kts
@@ -8,6 +8,7 @@ dependencyResolutionManagement {
     repositories {
         gradlePluginPortal()
 
-        maven("https://repo.crazycrew.us/api/")
+        maven("https://repo.crazycrew.us/first-party/")
+        maven("https://repo.crazycrew.us/third-party/")
     }
 }
