+++ buildSrc/src/main/kotlin/root-plugin.gradle.kts
@@ -4,15 +4,9 @@ plugins {
 }
 
 repositories {
-    maven("https://repo.extendedclip.com/content/repositories/placeholderapi/")
+    maven("https://repo.crazycrew.us/first-party/")
 
-    maven("https://repo.codemc.org/repository/maven-public/")
-
-    maven("https://repo.aikar.co/content/groups/aikar/")
-
-    maven("https://repo.triumphteam.dev/snapshots/")
-
-    maven("https://repo.crazycrew.us/api/")
+    maven("https://repo.crazycrew.us/third-party/")
 
     maven("https://jitpack.io/")
