+++ gradle.properties
@@ -4,5 +4,5 @@ org.gradle.warning.mode = all
 
 name = CrazyEnchantments
 group = com.badbones69.crazyenchantments
-version = 1.9.9.3
+version = 1.9.9.4
 description = Adds over 80 unique enchantments to your server and more!
