+++ settings.gradle.kts
@@ -8,7 +8,9 @@ pluginManagement {
         mavenCentral()
 
         maven("https://repo.papermc.io/repository/maven-public/")
-        maven("https://repo.crazycrew.us/api/")
+
+        maven("https://repo.crazycrew.us/first-party/")
+        maven("https://repo.crazycrew.us/third-party/")
     }
 }
